+++ tests/content_origin/test_repos.py
@@ -4,13 +4,31 @@ from rpm_lockfile.content_origin import Repo
 from rpm_lockfile.content_origin.repos import RepoOrigin
 
 
-def test_collect_simple(tmpdir):
+def test_collect_simple_baseurl(tmpdir):
     baseurl = "https://example.com/repo"
     config = [{"repoid": "a", "baseurl": baseurl}]
     origin = RepoOrigin(tmpdir)
     repos = list(origin.collect(config))
 
-    assert repos == [Repo(repoid="a", baseurl=baseurl)]
+    assert repos == [Repo(repoid="a", kwargs={"baseurl": [baseurl]})]
+
+
+def test_collect_simple_metalink(tmpdir):
+    url = "https://example.com/repo"
+    config = [{"repoid": "a", "metalink": url}]
+    origin = RepoOrigin(tmpdir)
+    repos = list(origin.collect(config))
+
+    assert repos == [Repo(repoid="a", kwargs={"metalink": url})]
+
+
+def test_collect_simple_mirrorlist(tmpdir):
+    url = "https://example.com/repo"
+    config = [{"repoid": "a", "mirrorlist": url}]
+    origin = RepoOrigin(tmpdir)
+    repos = list(origin.collect(config))
+
+    assert repos == [Repo(repoid="a", kwargs={"mirrorlist": url})]
 
 
 def fake_get_labels(obj, config_dir):
@@ -25,7 +43,9 @@ def fake_get_labels(obj, config_dir):
 TEMPLATE_CONFIG = {
     "repoid": "a", "baseurl": "https://example.com/{architecture}/repo"
 }
-EXPANDED_REPO = Repo(repoid="a", baseurl="https://example.com/x86_64/repo")
+EXPANDED_REPO = Repo(
+    repoid="a", kwargs={"baseurl": ["https://example.com/x86_64/repo"]}
+)
 
 
 def test_collect_with_vars_from_image(tmpdir):